+++ EntityLayer/Dtos/UpdateBusinessAdminDto.cs
@@ -8,8 +8,8 @@ namespace EntityLayer.Dtos
 {
     public class UpdateBusinessAdminDto
     {
-        public int Id { get; set; }
-        public int DeliveryAgentId { get; set; }
+        public long Id { get; set; }
+        public long DeliveryAgentId { get; set; }
         enum DeliveryAgentStatus
         {
             Availale,
